test/9_RMQ_FastAPI/UserManagement/requirements.txt
fastapi
uvicorn
pika
psycopg2-binary
